# crypto/secp256k1: enable 128-bit int code and endomorphism optimization